+++ tests/test_fixtures.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 import pytest
 from flask import request
 from flask import url_for
@@ -21,40 +20,24 @@ class TestFixtures:
         assert request_ctx.app is app
 
     def test_request_ctx_is_kept_around(self, client):
-        client.get(url_for("index"), headers=[("X-Something", "42")])
-        assert request.headers["X-Something"] == "42"
+        res = client.get(url_for("index"), headers=[("X-Something", "42")])
+        """In werkzeug 2.0.0 the test Client provides a new attribute 'request'
+        in the response class wich holds a reference to the request object that
+        produced the respective response, making instrospection easier"""
+        try:
+            assert res.request.headers["X-Something"] == "42"
+        except AttributeError:
+            """This is the conventional (pre 2.0.0) way of reaching the
+            request object, using flask.request global."""
+            assert request.headers["X-Something"] == "42"
 
+    def test_accept_mimetype(self, accept_mimetype):
+        mimestrings = [[("Accept", "application/json")], [("Accept", "text/html")]]
+        assert accept_mimetype in mimestrings
 
-class TestJSONResponse:
-    def test_json_response(self, client, accept_json):
-        res = client.get(url_for("ping"), headers=accept_json)
-        assert res.json == {"ping": "pong"}
-
-    def test_json_response_compare_to_status_code(self, client, accept_json):
-        assert client.get(url_for("ping"), headers=accept_json) == 200
-        assert client.get("fake-route", headers=accept_json) == 404
-        assert client.get("fake-route", headers=accept_json) != "404"
-        res = client.get(url_for("ping"), headers=accept_json)
-        assert res == res
-
-    def test_mismatching_eq_comparison(self, client, accept_json):
-        with pytest.raises(AssertionError, match=r"Mismatch in status code"):
-            assert client.get("fake-route", headers=accept_json) == 200
-        with pytest.raises(AssertionError, match=r"404 NOT FOUND"):
-            assert client.get("fake-route", headers=accept_json) == "200"
-
-    def test_dont_rewrite_existing_implementation(self, app, accept_json):
-        class MyResponse(app.response_class):
-            @property
-            def json(self):
-                """What is the meaning of life, the universe and everything?"""
-                return 42
-
-        app.response_class = MyResponse
-        client = app.test_client()
-
-        res = client.get(url_for("ping"), headers=accept_json)
-        assert res.json == 42
+    def test_accept_any(self, accept_any):
+        mimestrings = [[("Accept", "*")], [("Accept", "*/*")]]
+        assert accept_any in mimestrings
 
 
 @pytest.mark.usefixtures("client_class")